+++ NEWS.md
@@ -1,8 +1,13 @@
 # BuzzardsBay 0.1.0.9029
 
+Changes to `stitch_site`:
 - Quote strings when writing Archive and WPP files, as comments sometimes include
 commas, which are a no-no in CSVs.
-- 
+- Always use blank, not #N/A in CSVs
+- Reject any rows where Cal = 1 unless Gen_QC is 11 or 12
+
+
+
 
 # BuzzardsBay 0.1.0.9028